Move import process to middleware.
